+++ LRM Rust/src/pipeline/video.rs
@@ -5,11 +5,25 @@
 
 
 
-// use crate::pipeline::{tracker::LipTrackerConfig, FramesBuffer};
-// use opencv::{
-//     core::{AlgorithmHint, Mat, MatTraitConst, MatTraitConstManual},
-//     imgproc,
-//     videoio::{VideoCapture, VideoCaptureTrait, CAP_ANY},
-// };
-// use std::{error::Error, path::Path};
-// use crate::pipeline::tracker::LipTracker;
+use crate::pipeline::{
+    tracker::LipTrackerBackend,
+    FramesBuffer,
+};
+use opencv::{
+    core::{
+        AlgorithmHint,
+        Mat,
+        MatTraitConst,
+        MatTraitConstManual,
+    },
+    imgproc,
+    videoio::{
+        VideoCapture,
+        VideoCaptureTrait,
+        CAP_ANY,
+    },
+};
+use std::{error::Error, path::Path};
+
+
+